Jacobian mostly complete
Debugged most of the residual. Now all the values going into it are
correct and the output is correct (granted it’s mostly zeros at first,
so hard to check). There’s a few nan’s to check out.

Also finished the jacobian. All that is needed is to perform the edge
checks. if x==Nx-2, or Nx-1. etc.