Data Structure Update
Updated the SimState data structure.

Previously it stored arrays for c, phi, and alpha.

Now it stores: vector of all unknowns.
index sets for c, phi, and alpha.
subvectors that are just pointers to the big vector.
array versions of those subverters.

Now, we can update variables directly. Then use the array to refer to
the entries nicely. This causes a new issue to be mindful of.

WHENEVER STATE_VARS CHANGES YOU MUST CALL RESTORE_SUBARRAY AND THEN
EXTRACT_SUBARRAY.

This ensures data is up to date and safe.