Multigrid inclusion
The multigrid branch for whatever reason was at least twice as slow as
this branch.

Simplest fix was to just copy and paste the multigrid parts into this
branch. This gives us a method for doing multigrid in this now.
Currently it is not set to use multigrid since multigrid solving
appears to be about 50% slower than using ILU.