Transition to single state vector support
Update state variables to be properly saved and passed around from
their vector form.

Included a new data struct “AppCtx” to pass into the user context space
in the SNES package of Petsc. Everything is still the same but shifts
the variables to be saved in “user” for the most part with updates
being done through a Vec called current_state.