Fix stack overflow error. Also (possibly) fixed stability error
Fixed the one malloc I forgot to do. (also aligned the other mallocs to
be single lines)

Removing my set of the KSP tols. As well as the unpreconditioned norm
apparently made it more stable. (mainly the unpreconditioned part).

Also, with this setup (and for larger values of Nx). Multigrid with
richardson+sor is fastest.