FIXED
got the adaptive routine fixed.

The main problem seemed to be in not saving the past gating variable
values, so they updated and unuqdated in strange ways.

Added a new past and current storage bins for each. We pull the grid
from the past, and save the middle point to the current one. So now
there is an identical treatment for c,phi,alpha and all the gating
variables.