Grid solve as an initial guess maker
This version of code will take the result of the grid solver as an
initial guess for the SNESSolve.

Commenting out Update_Solve leads to the old solver, commenting out
SNESSolve leads to just using the grid solver.

Currently, this works fine for choosing one or the other. But something
funny happens with data structures when you use both of them.