Separate volume solve works. Also, predictor option
Made a new index function, Ind_2 that is always used with the grid
solve. The grid solve will always solve with alpha. Now you can solve
the bigger problem with alpha separate and it will work.

Also included a predictor option. With this enable the grid solve will
be used as a predictor. With it off the grid solve will not be used at
all. Instead we solve using the full jacobian iteration.