Predictor threshhold
Implemented a threshold on the predictor update. A point will only do
an update if it is above a given threshold. This at first glance
appears to keep the 2 Newton iteration behavior that was nice about the
previous version. But reduces the time sink considerably.

Before the Predictor would take about as long as the full solve. So it
doubled the time. Now it only takes 3% of the solve time.

Previously for a 64x64 problem ran for 99 time steps. It would run the
Grid_solve ~4000 times each time step. Now it runs ~120 times per time
step.