SPIRAL
Accidentally deleted the whole stem+circle at the critical time to
depolarize the whole circle quickly, but it spreads through. Triggering
a wave at the far side first. Creating a spiral near the top of the
domain.

Added back in what I meant to be looking at. But left the spiral as a
commented out section.