Start from given file
Added capability to start from initial values given in the save_csd.txt
file.

The way this is meant to be used is that it will save the data after a
simulation ends. The simulation can then be restarted. This assumes
that all parameters are kept the same as we only saving c,phi,alpha and
gating variables.

One thing to keep in mind is that excitation is based on time. Right
now you need to manually turn that off.