Read Only error fix
When compiling petsc in debugging mode. SNESSolve would throw an error
with the vectors being locked in read mode. Apparently switching the
extract_subarray and restore_subarray calls to use read only vectors in
this part fixes it.

However, we don’t need to always do read only (for initializing and
grid solving). So we needed two more functions.